# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 808 All Purpose Cleaners Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /labor-entries:
    get:
      operationId: listLaborEntries
      tags: [labor]
      summary: List labor entries
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of labor entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LaborEntry"
    post:
      operationId: createLaborEntry
      tags: [labor]
      summary: Create a labor entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLaborEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaborEntry"
  /labor-entries/{id}:
    patch:
      operationId: updateLaborEntry
      tags: [labor]
      summary: Update a labor entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLaborEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaborEntry"
    delete:
      operationId: deleteLaborEntry
      tags: [labor]
      summary: Delete a labor entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /followups:
    get:
      operationId: listFollowups
      tags: [followups]
      summary: List followups
      responses:
        "200":
          description: List of followups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Followup"
    post:
      operationId: createFollowup
      tags: [followups]
      summary: Create a followup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Followup"
  /followups/{id}:
    patch:
      operationId: updateFollowup
      tags: [followups]
      summary: Update a followup
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFollowupBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Followup"
    delete:
      operationId: deleteFollowup
      tags: [followups]
      summary: Delete a followup
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        hourlyRate:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, role, hourlyRate, status, createdAt, updatedAt]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        hourlyRate:
          type: string
        status:
          type: string
      required: [name]
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        hourlyRate:
          type: string
        status:
          type: string
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        amount:
          type: string
        tax:
          type: string
        total:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, amount, tax, total, status, createdAt, updatedAt]
    CreateInvoiceBody:
      type: object
      properties:
        bookingId:
          type: integer
        clientId:
          type: integer
        invoiceNumber:
          type: string
        amount:
          type: string
        tax:
          type: string
        total:
          type: string
        status:
          type: string
        dueDate:
          type: string
        paidDate:
          type: string
        description:
          type: string
        clientName:
          type: string
      required: [invoiceNumber, amount, total]
    UpdateInvoiceBody:
      type: object
      properties:
        status:
          type: string
        paidDate:
          type: ["string", "null"]
        amount:
          type: string
        tax:
          type: string
        total:
          type: string
        dueDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: string
        amount:
          type: string
        date:
          type: string
        vendor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, category, description, amount, date, createdAt, updatedAt]
    CreateExpenseBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: string
        date:
          type: string
        vendor:
          type: string
        notes:
          type: string
      required: [category, description, amount, date]
    UpdateExpenseBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: string
        date:
          type: string
        vendor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    LaborEntry:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        bookingId:
          type: ["integer", "null"]
        date:
          type: string
        hoursWorked:
          type: string
        hourlyRate:
          type: string
        totalPay:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, employeeId, date, hoursWorked, hourlyRate, totalPay, createdAt, updatedAt]
    CreateLaborEntryBody:
      type: object
      properties:
        employeeId:
          type: integer
        bookingId:
          type: integer
        date:
          type: string
        hoursWorked:
          type: string
        hourlyRate:
          type: string
        totalPay:
          type: string
        description:
          type: string
      required: [employeeId, date, hoursWorked, hourlyRate, totalPay]
    UpdateLaborEntryBody:
      type: object
      properties:
        date:
          type: string
        hoursWorked:
          type: string
        hourlyRate:
          type: string
        totalPay:
          type: string
        description:
          type: ["string", "null"]
...
    Followup:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        reason:
          type: string
        dueDate:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientName, reason, dueDate, status, createdAt, updatedAt]
    CreateFollowupBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        reason:
          type: string
        dueDate:
          type: string
        status:
          type: string
        notes:
          type: string
      required: [clientName, reason, dueDate]
    UpdateFollowupBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        reason:
          type: string
        dueDate:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
...