# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 808 All Purpose Cleaners Business Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceBody {
  bookingId?: number;
  clientId?: number;
  invoiceNumber: string;
  amount: string;
  tax?: string;
  total: string;
  status?: string;
  dueDate?: string;
  paidDate?: string;
  description?: string;
  clientName?: string;
}