# lib/api-zod/src/generated/types/createLaborEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 808 All Purpose Cleaners Business Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLaborEntryBody {
  employeeId: number;
  bookingId?: number;
  date: string;
  hoursWorked: string;
  hourlyRate: string;
  totalPay: string;
  description?: string;
}