# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 808 All Purpose Cleaners Business Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  category: string;
  description: string;
  amount: string;
  date: string;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}